requirements.txt
openai
pandas
fastapi
uvicorn
google-generativeai
dotenvs